PythonRag/pyproject.toml
[project]
name = "python-rag-app"
version = "1.0.0"
description = "A Python RAG application using Azure OpenAI and Azure AI Search"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-ai-openai>=2.0.0",
    "azure-identity>=1.13.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["."]

[tool.setuptools]
packages = ["python_rag_app", "python_rag_app.services"]
